ENH: Add support for 16 and 32-bit radiograph images
Prior to these changes, the supported CUDA texture types (and by extension
radiograph images) was hard-coded to 8-bit (unsigned char) and managed as
a static global variable.

Since texture references needed to be known at compile time and the current
type was hard-coded to 8-bit, the use of them prevented from being able to
easily support 8, 16, and 32 bit radiograph images known at runtime.

To address this, this commit removes the use of `cudaBindTextureToArray`
(Deprecated since CUDA 11.3 and removed in CUDA 12.x)
and switch to using texture object instead by introducing a convenience helper
function `createTextureObjectFromArray`.

Approach implemented here is based of recommendation the post titled
"Kepler Texture Objects Improve Performance and Flexibility" published
in 2013.
See https://developer.nvidia.com/blog/cuda-pro-tip-kepler-texture-objects-improve-performance-and-flexibility/

For reference, before this changes, the error message reported at build time
while attempting to use autoscoper built against CUDA 12.x was:

```
3 Errors under file path `...\libautoscoper\src\gpu\cuda\RadRenderer_kernels.cu:
- no instance of overloaded function "tex2D" matches the argument list
- identifier "cudaBindTextureToArray" is undefined
- texture is not a template
```

+++ libautoscoper/src/gpu/cuda/Merger_kernels.cu
@@ -60,8 +60,8 @@ void merge(float* src1,
 {
     // Calculate the block and grid sizes.
     dim3 blockDim(32, 32);
-    dim3 gridDim((width+blockDim.x-1)/blockDim.x,
-                 (height+blockDim.y-1)/blockDim.y);
+    dim3 gridDim(((unsigned int)width+blockDim.x-1)/blockDim.x,
+                 ((unsigned int)height+blockDim.y-1)/blockDim.y);
 
     // Call the kernel
     merge_kernel<<<gridDim, blockDim>>>(src1,src2,dest,width,height);
